refactor: Parameter-related processing cut out to file
